Assigments/02_Assigment_02/01_Basic FAQ Agent/pyproject.toml
[project]
name = "01-basic-faq-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
